// src/app/shared/model/axis.enum.ts
export enum Weekday {
  Sunday = 'Sunday',
  Monday = 'Monday',
  Tuesday = 'Tuesday',
  Wednesday = 'Wednesday',
  Thursday = 'Thursday',
  Friday = 'Friday',
  Saturday = 'Saturday',
}

export enum Hour {
  AM12 = '12AM',
  AM1 = '1AM',
  AM2 = '2AM',
  AM3 = '3AM',
  AM4 = '4AM',
  AM5 = '5AM',
  AM6 = '6AM',
  AM7 = '7AM',
  AM8 = '8AM',
  AM9 = '9AM',
  AM10 = '10AM',
  AM11 = '11AM',
  PM12 = '12PM',
  PM1 = '1PM',
  PM2 = '2PM',
  PM3 = '3PM',
  PM4 = '4PM',
  PM5 = '5PM',
  PM6 = '6PM',
  PM7 = '7PM',
  PM8 = '8PM',
  PM9 = '9PM',
  PM10 = '10PM',
  PM11 = '11PM',
}